Extract error module in model
