+++ dsp-meta/src/domain/model/mod.rs
@@ -1,2 +1,3 @@
 pub mod draft_model;
+pub mod error;
 pub mod json_schema_validator;
